refactor: menu_group 패키지 menu로 이동
