+++ src/main/java/kitchenpos/menu/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu_group;
+package kitchenpos.menu;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
