+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 import kitchenpos.menu.Menu;
 import kitchenpos.menu.MenuProduct;
+import kitchenpos.menu.persistence.MenuGroupRepository;
 import kitchenpos.menu.persistence.MenuRepository;
 import kitchenpos.menu.request.MenuCreateRequest;
 import kitchenpos.menu.request.MenuProductCreateRequest;
-import kitchenpos.menu_group.persistence.MenuGroupRepository;
 import kitchenpos.product.Product;
 import kitchenpos.product.persistence.ProductRepository;
 import org.springframework.stereotype.Service;
