+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -7,12 +7,12 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.Menu;
+import kitchenpos.menu.MenuGroup;
 import kitchenpos.menu.MenuProduct;
+import kitchenpos.menu.persistence.MenuGroupRepository;
 import kitchenpos.menu.persistence.MenuRepository;
 import kitchenpos.menu.request.MenuCreateRequest;
 import kitchenpos.menu.request.MenuProductCreateRequest;
-import kitchenpos.menu_group.MenuGroup;
-import kitchenpos.menu_group.persistence.MenuGroupRepository;
 import kitchenpos.product.Product;
 import kitchenpos.product.persistence.ProductRepository;
 import kitchenpos.support.ServiceTest;
