+++ src/test/java/kitchenpos/table_group/application/TableGroupServiceTest.java
@@ -9,10 +9,10 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.menu.Menu;
+import kitchenpos.menu.MenuGroup;
 import kitchenpos.menu.MenuProduct;
+import kitchenpos.menu.persistence.MenuGroupRepository;
 import kitchenpos.menu.persistence.MenuRepository;
-import kitchenpos.menu_group.MenuGroup;
-import kitchenpos.menu_group.persistence.MenuGroupRepository;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderStatus;
